{"schema":"decodex/commit/1","summary":"Consolidate over-split Swift app helpers","authority":"manual"}
