--- apps/decodex-app/Sources/DecodexApp/AccountControl.swift
@@ -1,11 +1,0 @@
-import Foundation
-
-struct AccountControl: Decodable {
-	let mode: String
-	let accountSelector: String?
-
-	enum CodingKeys: String, CodingKey {
-		case mode
-		case accountSelector = "account_selector"
-	}
-}
