--- apps/decodex-app/Sources/DecodexApp/AccountPanelPrivacy.swift
@@ -1,4 +1,0 @@
-enum AccountPrivacy {
-	static let hiddenValue = "hidden"
-	static let visibleValue = "visible"
-}
