+++ apps/decodex-app/Sources/DecodexApp/AccountStore.swift
@@ -96,3 +96,13 @@ final class AccountStore: ObservableObject {
 		notice = nil
 	}
 }
+
+struct CodexFastModeResponse: Decodable, Equatable {
+	let codexConfigPath: String
+	let enabled: Bool
+
+	enum CodingKeys: String, CodingKey {
+		case codexConfigPath = "codex_config_path"
+		case enabled
+	}
+}
