--- apps/decodex-app/Sources/DecodexApp/CodexFastModeResponse.swift
@@ -1,11 +1,0 @@
-import Foundation
-
-struct CodexFastModeResponse: Decodable, Equatable {
-	let codexConfigPath: String
-	let enabled: Bool
-
-	enum CodingKeys: String, CodingKey {
-		case codexConfigPath = "codex_config_path"
-		case enabled
-	}
-}
