+++ apps/decodex-app/Sources/DecodexApp/DashboardWebSocketConnection.swift
@@ -1,5 +1,7 @@
 import Foundation
 import Network
+import Security
+
 
 actor DashboardWebSocketConnection {
 	let url: URL
@@ -64,5 +66,19 @@ actor DashboardWebSocketConnection {
 		]
 
 		return Data(lines.joined(separator: "\r\n").utf8)
+	}
+}
+
+extension DashboardWebSocketConnection {
+	func randomData(byteCount: Int) -> Data {
+		var bytes = [UInt8](repeating: 0, count: byteCount)
+		let status = SecRandomCopyBytes(kSecRandomDefault, bytes.count, &bytes)
+		if status != errSecSuccess {
+			for index in bytes.indices {
+				bytes[index] = UInt8.random(in: UInt8.min...UInt8.max)
+			}
+		}
+
+		return Data(bytes)
 	}
 }
