--- apps/decodex-app/Sources/DecodexApp/LoginFloatingPanel.swift
@@ -1,11 +1,0 @@
-import AppKit
-
-final class LoginFloatingPanel: NSPanel {
-	override var canBecomeKey: Bool {
-		true
-	}
-
-	override var canBecomeMain: Bool {
-		false
-	}
-}
