--- apps/decodex-app/Sources/DecodexApp/OperatorLaneReadoutSupport.swift
@@ -1,5 +1,0 @@
-import Foundation
-
-func rawPanelToken(_ value: String) -> String {
-	value.trimmingCharacters(in: .whitespacesAndNewlines)
-}
