--- apps/decodex-app/Sources/DecodexApp/PanelGlassSurfaceAPI.swift
@@ -1,22 +1,0 @@
-import SwiftUI
-
-enum GlassSurfaceDepth {
-	case panel
-	case section
-	case row
-	case control
-}
-
-extension View {
-	func modernGlassSurface(
-		cornerRadius: CGFloat,
-		depth: GlassSurfaceDepth = .section
-	) -> some View {
-		modifier(
-			ModernGlassSurfaceModifier(
-				cornerRadius: cornerRadius,
-				depth: depth
-			)
-		)
-	}
-}
