--- apps/decodex-app/Sources/DecodexApp/PanelWindowSizingProbeView.swift
@@ -1,10 +1,0 @@
-import AppKit
-
-final class PanelWindowSizingProbeView: NSView {
-	var didMoveToWindow: ((PanelWindowSizingProbeView) -> Void)?
-
-	override func viewDidMoveToWindow() {
-		super.viewDidMoveToWindow()
-		didMoveToWindow?(self)
-	}
-}
